# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Quizzz / ExamPro API
  version: 1.0.0
  description: |
    REST API for educational exams: auth, questions (with optional `image_urls`), exams, sessions, and media upload to Cloudinary.

    **Base URL:** `/api/v1` (e.g. `http://localhost:8080/api/v1`). Health: `GET /health`.

    **Auth:** most routes require `Authorization: Bearer <JWT>` from `POST /api/v1/auth/login`.
servers:
  - url: /
    description: Server root (e.g. http://localhost:8080)
...
paths:
...
  /api/v1/sessions/{id}:
    get:
      tags: [sessions]
      summary: Get session state
      security:
        - bearerAuth: []
    put:
      tags: [sessions]
      summary: Submit answer or finish (see handler)
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...